requirements.txt
pandas
numpy
scikit-learn
mlflow
dvc
pyyaml
joblib
matplotlib